+++ src/Command/DeployTasksCreateSchemaCommand.php
@@ -46,10 +46,9 @@ final class DeployTasksCreateSchemaCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
-        $sql = $this->storage->getCreateTableSql();
 
         if (true === $input->getOption('dump-sql')) {
-            $output->writeln($sql.';');
+            $output->writeln($this->storage->getCreateTableSql().';');
 
             return Command::SUCCESS;
         }